e2e: move tolerations tests to their own file
reduce the load on already big `configuration.go`
and make room for upcoming tests
